Install play.py script and bump release to 0.0.2
